Gen ID Algo
Problem Statement: 
A Store has released a new doll and public demand for it is huge. To prevent scalpers, Brand Marketing has decided to create a raffle system. The raffle system requires customers to register and receive a Raffle ID. The Raffle ID is the user registration table's primary key ID. After demoing the project to the Brand Marketing team, they quickly realized that a numeric registration ID was not ideal. Instead, Brand Marketing team initiated a change request where instead of showing a raffle number based off the customer registration sequence, it would instead show a "Short Code". The pattern for the Short Code is 3 letter, 3 numbers:
<LETTER><LETTER><LETTER><NUMBER><NUMBER><NUMBER>
Sample: ABC123
In order to minimize the code changes already in place, the customer's registration would still
create a primary key, but shown to the customer would be the Short Code instead.
Examples of the primary key id to Short Code:
users.id[1] is shown as AAA001
users.id[2] is shown as AAA002
users.id[3] is shown as AAA003
...
users.id[1000] is shown as AAB000
users.id[1001] is shown as AAB001
users.id[1002] is shown as AAB002
....
users.id[345234] is shown as ANH234
users.id[345235] is shown as ANH235
....
users.id[17575999] is shown as ZZZ999

Note: There is a limit of 17,575,999 possible combinations with a 3 letter & 3 number format.
Only numbers between 0 & 17,575,999 are accepted.